fzz: Move to engine and rename to packager
